+++ geolocator/test/geolocator_test.dart
@@ -12,10 +12,14 @@ Position get mockPosition => Position(
       isUtc: true,
     ),
     altitude: 3000.0,
+    satelliteCount: 3.0,
+    satellitesUsedInFix: 2.0,
     accuracy: 0.0,
     heading: 0.0,
     speed: 0.0,
-    speedAccuracy: 0.0);
+    speedAccuracy: 0.0,
+    altitudeAccuracy: 0.0,
+    headingAccuracy: 0.0);
 
 void main() {
   group('Geolocator', () {